NET-67: move claims to separate files
