+++ consensus/packets/claim_serializers.go
@@ -1,0 +1,273 @@
+/*
+ *    Copyright 2018 Insolar
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package packets
+
+import (
+	"encoding/binary"
+	"io"
+
+	"github.com/pkg/errors"
+)
+
+// claims auxiliar constants
+const (
+	claimTypeShift      = 10
+	claimHeaderTypeMask = 0xfc00
+
+	//	claimHeaderLengthMask = 0x3ff
+)
+
+func extractClaimTypeFromHeader(claimHeader uint16) uint8 {
+	return uint8((claimHeader & claimHeaderTypeMask) >> claimTypeShift)
+}
+
+// func extractClaimLengthFromHeader(claimHeader uint16) uint16 {
+// 	return claimHeader & claimHeaderLengthMask
+// }
+
+func makeClaimHeader(claim ReferendumClaim) uint16 {
+	// TODO: we don't need length
+	var result = 0 // claim.Length()
+	result |= (uint16(claim.Type()) << claimTypeShift)
+
+	return result
+}
+
+// Deserialize implements interface method
+func (nb *NodeBroadcast) Deserialize(data io.Reader) error {
+	err := binary.Read(data, defaultByteOrder, &nb.EmergencyLevel)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeBroadcast.Deserialize ] Can't read EmergencyLevel")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &nb.length)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeBroadcast.Deserialize ] Can't read length")
+	}
+
+	return nil
+}
+
+// Serialize implements interface method
+func (nb *NodeBroadcast) Serialize() ([]byte, error) {
+	result := allocateBuffer(64)
+	err := binary.Write(result, defaultByteOrder, nb.EmergencyLevel)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeBroadcast.Serialize ] Can't write EmergencyLevel")
+	}
+
+	err = binary.Write(result, defaultByteOrder, nb.length)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeBroadcast.Serialize ] Can't write length")
+	}
+
+	return result.Bytes(), nil
+}
+
+// Deserialize implements interface method
+func (cpa *CapabilityPoolingAndActivation) Deserialize(data io.Reader) error {
+	err := binary.Read(data, defaultByteOrder, &cpa.PollingFlags)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeBroadcast.Deserialize ] Can't read PollingFlags")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &cpa.CapabilityType)
+	if err != nil {
+		return errors.Wrap(err, "[ CapabilityPoolingAndActivation.Deserialize ] Can't read CapabilityType")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &cpa.CapabilityRef)
+	if err != nil {
+		return errors.Wrap(err, "[ CapabilityPoolingAndActivation.Deserialize ] Can't read CapabilityRef")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &cpa.length)
+	if err != nil {
+		return errors.Wrap(err, "[ CapabilityPoolingAndActivation.Deserialize ] Can't read length")
+	}
+
+	return nil
+}
+
+// Serialize implements interface method
+func (cpa *CapabilityPoolingAndActivation) Serialize() ([]byte, error) {
+	result := allocateBuffer(68)
+	err := binary.Write(result, defaultByteOrder, cpa.PollingFlags)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ CapabilityPoolingAndActivation.Serialize ] Can't write PollingFlags")
+	}
+
+	err = binary.Write(result, defaultByteOrder, cpa.CapabilityType)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ CapabilityPoolingAndActivation.Serialize ] Can't write CapabilityType")
+	}
+
+	err = binary.Write(result, defaultByteOrder, cpa.CapabilityRef)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ CapabilityPoolingAndActivation.Serialize ] Can't write CapabilityRef")
+	}
+
+	return result.Bytes(), nil
+}
+
+// Deserialize implements interface method
+func (nvb *NodeViolationBlame) Deserialize(data io.Reader) error {
+	err := binary.Read(data, defaultByteOrder, &nvb.BlameNodeID)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeViolationBlame.Deserialize ] Can't read BlameNodeID")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &nvb.TypeViolation)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeViolationBlame.Deserialize ] Can't read TypeViolation")
+	}
+
+	err = binary.Read(data, defaultByteOrder, nvb.Type())
+	if err != nil {
+		return errors.Wrap(err, "[ NodeViolationBlame.Deserialize ] Can't read claimType")
+	}
+
+	return nil
+}
+
+// Serialize implements interface method
+func (nvb *NodeViolationBlame) Serialize() ([]byte, error) {
+	result := allocateBuffer(5)
+	err := binary.Write(result, defaultByteOrder, nvb.BlameNodeID)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeViolationBlame.Serialize ] Can't write BlameNodeID")
+	}
+
+	err = binary.Write(result, defaultByteOrder, nvb.TypeViolation)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeViolationBlame.Serialize ] Can't write TypeViolation")
+	}
+
+	return result.Bytes(), nil
+}
+
+// Deserialize implements interface method
+func (njc *NodeJoinClaim) Deserialize(data io.Reader) error {
+	err := binary.Read(data, defaultByteOrder, &njc.NodeID)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeJoinClaim.Deserialize ] Can't read NodeID")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &njc.RelayNodeID)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeJoinClaim.Deserialize ] Can't read RelayNodeID")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &njc.ProtocolVersionAndFlags)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeJoinClaim.Deserialize ] Can't read ProtocolVersionAndFlags")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &njc.JoinsAfter)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeJoinClaim.Deserialize ] Can't read JoinsAfter")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &njc.NodeRoleRecID)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeJoinClaim.Deserialize ] Can't read NodeRoleRecID")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &njc.NodeRef)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeJoinClaim.Deserialize ] Can't read NodeRef")
+	}
+
+	err = binary.Read(data, defaultByteOrder, &njc.NodePK)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeJoinClaim.Deserialize ] Can't read NodePK")
+	}
+
+	// err = binary.Read(data, defaultByteOrder, &njc.length)
+	// if err != nil {
+	// 	return errors.Wrap(err, "[ NodeJoinClaim.Deserialize ] Can't read length")
+	// }
+
+	return nil
+}
+
+// Serialize implements interface method
+func (njc *NodeJoinClaim) Serialize() ([]byte, error) {
+	result := allocateBuffer(1024)
+	err := binary.Write(result, defaultByteOrder, njc.NodeID)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeJoinClaim.Serialize ] Can't write NodeID")
+	}
+
+	err = binary.Write(result, defaultByteOrder, njc.RelayNodeID)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeJoinClaim.Serialize ] Can't write RelayNodeID")
+	}
+
+	err = binary.Write(result, defaultByteOrder, njc.ProtocolVersionAndFlags)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeJoinClaim.Serialize ] Can't write ProtocolVersionAndFlags")
+	}
+
+	err = binary.Write(result, defaultByteOrder, njc.JoinsAfter)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeJoinClaim.Serialize ] Can't write JoinsAfter")
+	}
+
+	err = binary.Write(result, defaultByteOrder, njc.NodeRoleRecID)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeJoinClaim.Serialize ] Can't write NodeRoleRecID")
+	}
+
+	err = binary.Write(result, defaultByteOrder, njc.NodeRef)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeJoinClaim.Serialize ] Can't write NodeRef")
+	}
+
+	err = binary.Write(result, defaultByteOrder, njc.NodePK)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeJoinClaim.Serialize ] Can't write NodePK")
+	}
+
+	// err = binary.Write(result, defaultByteOrder, njc.length)
+	// if err != nil {
+	// 	return nil, errors.Wrap(err, "[ NodeJoinClaim.Serialize ] Can't write length")
+	// }
+
+	return result.Bytes(), nil
+}
+
+// Deserialize implements interface method
+func (nlc *NodeLeaveClaim) Deserialize(data io.Reader) error {
+	err := binary.Read(data, defaultByteOrder, &nlc.length)
+	if err != nil {
+		return errors.Wrap(err, "[ NodeLeaveClaim.Deserialize ] Can't read length")
+	}
+
+	return nil
+}
+
+// Serialize implements interface method
+func (nlc *NodeLeaveClaim) Serialize() ([]byte, error) {
+	result := allocateBuffer(64)
+	err := binary.Write(result, defaultByteOrder, nlc.length)
+	if err != nil {
+		return nil, errors.Wrap(err, "[ NodeLeaveClaim.Serialize ] Can't write length")
+	}
+
+	return result.Bytes(), nil
+}
